refactor: unify Rust code idioms and style
Reasons:
- Improve code readability and maintainability.
- Follow modern Rust programming best practices.

Changes:
- Simplify `Ok(())` return statements.
- Use the iterator method `find` instead of manual loop-based searches.
- Optimize `HashMap` insertion logic using `entry().or_insert_with()` and `or_default()`.
- Improve error handling by using `is_err()`.
- Adopt struct field initialization shorthand.
- Implement the `Default` trait for multiple structs.
- Use the `div_ceil` function for integer ceiling division.
- Optimize trait object references by directly returning `&dyn Trait`.
- Add `#[allow(clippy::...)` attributes to suppress specific lint warnings.

+++ src-tauri/src/resource/controller/skydimo_serial/config.rs
@@ -582,17 +582,17 @@ fn build_matrix_for_config(config: &SkydimoModelConfig) -> Option<SkydimoMatrixL
 
     let (height, width) = match config.layout {
         Perimeter4 => {
-            let h = (z1.max(z3) + 2) as usize;
-            let w = (z2.max(z4) + 2) as usize;
+            let h = z1.max(z3) + 2;
+            let w = z2.max(z4) + 2;
             (h, w)
         }
         Perimeter3 => {
-            let h = (z1.max(z3) + 1) as usize;
-            let w = (z2 + 2) as usize;
+            let h = z1.max(z3) + 1;
+            let w = z2 + 2;
             (h, w)
         }
         Sides2 => {
-            let h = (z1.max(z2) + 2) as usize;
+            let h = z1.max(z2) + 2;
             let mut w_f = ((16.0 / 9.0) * (h as f64)).round() as isize;
             if w_f < 3 {
                 w_f = 3;
